Penambahan Class Baru & Pemindahan variabel beserta fungsinya - Large Class
Membuat Class Baru, yaitu :
1. Child
2. Salary
3. Spouse
4. Working

kemudian, memindahkan variabel  dan fungsi tertentu yang berada di Employee Class ke class yang baru dibuat.

Tujuannya untuk mengurangi baris line code yang cukup panjang agar lebih efektif
